+++ bin/ts-import-move.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+
+// This is just a wrapper script that loads the compiled version
+require('../dist/index.js');